+++ CHANGELOG.md
@@ -5,6 +5,13 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Change the `Interval`/`RequestedInterval` enum to include the new fields.
+  (FifteenSeconds,ThirtySeconds,OneMinutes)
+
 ## [0.3.2] - 2024-01-25
 
 ### Changed